Reorganized the OperationHelper
Added SyncOperationHelper and AsyncOperationHelper.
Improved readibility of the classes. Also makes refactoring methods
in the future simplier.

Finally, the refactoring identified methods that weren't used in
the operations at all. These have now been removed.
