+++ driver-core/src/main/com/mongodb/internal/operation/ListCollectionsOperation.java
@@ -61,16 +61,16 @@ import static com.mongodb.internal.operation.CommandOperationHelper.executeComma
 import static com.mongodb.internal.operation.CommandOperationHelper.isNamespaceError;
 import static com.mongodb.internal.operation.CommandOperationHelper.rethrowIfNotNamespaceError;
 import static com.mongodb.internal.operation.CursorHelper.getCursorDocumentFromBatchSize;
-import static com.mongodb.internal.operation.OperationHelper.AsyncCallableWithConnectionAndSource;
-import static com.mongodb.internal.operation.OperationHelper.CallableWithSource;
+import static com.mongodb.internal.operation.AsyncOperationHelper.AsyncCallableWithConnectionAndSource;
+import static com.mongodb.internal.operation.SyncOperationHelper.CallableWithSource;
 import static com.mongodb.internal.operation.OperationHelper.LOGGER;
-import static com.mongodb.internal.operation.OperationHelper.createEmptyAsyncBatchCursor;
-import static com.mongodb.internal.operation.OperationHelper.createEmptyBatchCursor;
-import static com.mongodb.internal.operation.OperationHelper.cursorDocumentToAsyncBatchCursor;
-import static com.mongodb.internal.operation.OperationHelper.cursorDocumentToBatchCursor;
-import static com.mongodb.internal.operation.OperationHelper.releasingCallback;
-import static com.mongodb.internal.operation.OperationHelper.withAsyncReadConnection;
-import static com.mongodb.internal.operation.OperationHelper.withReadConnectionSource;
+import static com.mongodb.internal.operation.AsyncOperationHelper.createEmptyAsyncBatchCursor;
+import static com.mongodb.internal.operation.SyncOperationHelper.createEmptyBatchCursor;
+import static com.mongodb.internal.operation.AsyncOperationHelper.cursorDocumentToAsyncBatchCursor;
+import static com.mongodb.internal.operation.SyncOperationHelper.cursorDocumentToBatchCursor;
+import static com.mongodb.internal.operation.AsyncOperationHelper.releasingCallback;
+import static com.mongodb.internal.operation.AsyncOperationHelper.withAsyncReadConnection;
+import static com.mongodb.internal.operation.SyncOperationHelper.withReadConnectionSource;
 import static com.mongodb.internal.operation.ServerVersionHelper.serverIsAtLeastVersionThreeDotZero;
 import static java.lang.String.format;
 import static java.util.Arrays.asList;
